new spellcasting system & refactor glyph handling
in this system, 'drawing' glyphs is the same as a kind of 'casting' as would be any specific spell casting action.

the system handles exposing effects, which are operated on by effect systems that render and display all effects.
it also handles blending between states, for example between drawing & casting states.
the system implements full separation of server and client logic, with utilities for passing events between them.